+++ rust/src/fundamental/context.rs
@@ -1014,6 +1014,8 @@ impl FundamentalContext {
             country: detail.market,
             name: detail.indicator_name,
             describe: detail.description,
+            periodicity: detail.frequence,
+            importance: detail.importance,
             ..Default::default()
         };